+++ main.cpp
@@ -1295,7 +1295,7 @@ int read_input(std::vector<source> &sources, char *fname, int maxzoom, int minzo
 		}
 
 		if (sources[source].file.size() > 4 && sources[source].file.substr(sources[source].file.size() - 4) == std::string(".csv")) {
-			long long layer_seq[CPUS];
+			long layer_seq[CPUS];
 			double dist_sums[CPUS];
 			size_t dist_counts[CPUS];
